extra/requirements.txt
torch
transformers
clang
python-dotenv
unidiff
pygithub
tqdm
scikit-learn 